LayoutTests/http/tests/site-isolation/datalist-cross-origin-iframe-expected.txt

Test that selecting a datalist option in a cross-origin iframe updates the input value when site isolation is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS selectedValue is "Apple"
PASS successfullyParsed is true

TEST COMPLETE
